Enable hostname verification on 1.0.2
